perf: NetId is now 2 bytes instead of 4
NetId is now a ushort instead of uint.
That means the max amount of networked objects is 65K
Note NetId can now recycled after a while.

BREAKING CHANGE: NetId is now a ushort instead of uint

+++ Assets/Mirror/Runtime/ClientObjectManager.cs
@@ -373,7 +373,7 @@ namespace Mirror
             ApplySpawnPayload(identity, msg);
         }
 
-        NetworkIdentity GetExistingObject(uint netid)
+        NetworkIdentity GetExistingObject(ushort netid)
         {
             client.Spawned.TryGetValue(netid, out NetworkIdentity localObject);
             return localObject;
@@ -446,7 +446,7 @@ namespace Mirror
             DestroyObject(msg.netId);
         }
 
-        void DestroyObject(uint netId)
+        void DestroyObject(ushort netId)
         {
             if (logger.LogEnabled()) logger.Log("ClientScene.OnObjDestroy netId:" + netId);
