+++ Assets/Mirror/Runtime/Messages.cs
@@ -34,7 +34,7 @@ namespace Mirror
     #region System Messages requried for code gen path
     public struct ServerRpcMessage
     {
-        public uint netId;
+        public ushort netId;
         public int componentIndex;
         public int functionHash;
 
@@ -54,7 +54,7 @@ namespace Mirror
 
     public struct RpcMessage
     {
-        public uint netId;
+        public ushort netId;
         public int componentIndex;
         public int functionHash;
         // the parameters for the Cmd function
@@ -69,7 +69,7 @@ namespace Mirror
         /// <summary>
         /// netId of new or existing object
         /// </summary>
-        public uint netId;
+        public ushort netId;
         /// <summary>
         /// Is the spawning object the local player. Sets ClientScene.localPlayer
         /// </summary>
@@ -108,17 +108,17 @@ namespace Mirror
 
     public struct ObjectDestroyMessage
     {
-        public uint netId;
+        public ushort netId;
     }
 
     public struct ObjectHideMessage
     {
-        public uint netId;
+        public ushort netId;
     }
 
     public struct UpdateVarsMessage
     {
-        public uint netId;
+        public ushort netId;
         // the serialized component data
         // -> ArraySegment to avoid unnecessary allocations
         public ArraySegment<byte> payload;
